requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
pandas==1.3.3
requests==2.26.0
scikit-learn==0.24.2
python-dotenv==0.19.0
flask-login==0.5.0
pytest==6.2.5
black==21.9b0
flake8==3.9.2